refactor(fetcher): move apiclient to it's own type
